+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/data/AttachedDatabaseInfo.kt
@@ -1,0 +1,6 @@
+package org.dbtools.room.data
+
+data class AttachedDatabaseInfo(
+    val name: String,
+    val file: String
+)