+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/ext/OkioFilesystemExt.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DuplicatedCode", "unused")
+
 package org.dbtools.room.ext
 
 import okio.FileSystem