+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/ext/SQLiteStatementExt.kt
@@ -19,3 +19,34 @@ fun SQLiteStatement.getColumnIndexOrThrow(name: String): Int {
     val availableColumns = List(getColumnCount()) { getColumnName(it) }.joinToString()
     throw IllegalArgumentException("Column '$name' does not exist. Available columns: [$availableColumns]")
 }
+
+/**
+ * Binds the provided arguments to the SQLiteStatement.
+ *
+ * @param args Variable number of arguments to bind to the statement.
+ * @throws IllegalArgumentException if an unsupported argument type is encountered.
+ */
+fun SQLiteStatement.bindArgs(args: List<Any?>) {
+    var index = 1
+    args.forEach { arg ->
+        when (arg) {
+            is String -> bindText(index, arg)
+            is Int -> bindInt(index, arg)
+            is Long -> bindLong(index, arg)
+            is Double -> bindDouble(index, arg)
+            is Float -> bindFloat(index, arg)
+            is Boolean -> bindBoolean(index, arg)
+            is ByteArray -> bindBlob(index, arg)
+            null -> bindNull(index)
+            else -> throw IllegalArgumentException("Unsupported argument type: ${arg::class}")
+        }
+        index++
+    }
+}
+
+fun SQLiteStatement.getTextOrNull(index: Int): String? = if (isNull(index)) null else getText(index)
+fun SQLiteStatement.getIntOrNull(index: Int): Int? = if (isNull(index)) null else getInt(index)
+fun SQLiteStatement.getLongOrNull(index: Int): Long? = if (isNull(index)) null else getLong(index)
+fun SQLiteStatement.getDoubleOrNull(index: Int): Double? = if (isNull(index)) null else getDouble(index)
+fun SQLiteStatement.getFloatOrNull(index: Int): Float? = if (isNull(index)) null else getFloat(index)
+fun SQLiteStatement.getBooleanOrNull(index: Int): Boolean? = if (isNull(index)) null else getBoolean(index)