+++ dbtools-room/src/commonMain/kotlin/org/dbtools/room/ext/TransactorExt.kt
@@ -3,11 +3,12 @@
 package org.dbtools.room.ext
 
 import androidx.room.Room
+import androidx.room.Transactor
+import androidx.room.execSQL
 import androidx.room.util.getColumnIndexOrThrow
-import androidx.sqlite.SQLiteConnection
 import androidx.sqlite.SQLiteStatement
-import androidx.sqlite.execSQL
 import co.touchlab.kermit.Logger
+import kotlinx.coroutines.runBlocking
 import okio.FileSystem
 import okio.Path
 import org.dbtools.room.DatabaseViewQuery
@@ -24,7 +25,7 @@ import kotlin.time.TimeSource.Monotonic.markNow
  * @return true if validation check is OK
  */
 @Suppress("NestedBlockDepth")
-fun SQLiteConnection.validateDatabase(tag: String = "", tableDataCountCheck: String? = null, allowZeroCount: Boolean = true): Boolean {
+suspend fun Transactor.validateDatabase(tag: String = "", tableDataCountCheck: String? = null, allowZeroCount: Boolean = true): Boolean {
     Logger.i { "Checking database integrity for [$tag]" }
     val mark = markNow()
     try {
@@ -36,7 +37,7 @@ fun SQLiteConnection.validateDatabase(tag: String = "", tableDataCountCheck: Str
 
         // make sure there is data in the database
         if (!tableDataCountCheck.isNullOrBlank()) {
-            prepare("SELECT count(1) FROM $tableDataCountCheck").use { statement ->
+            usePrepared("SELECT count(1) FROM $tableDataCountCheck") { statement ->
                 val count = if (statement.step()) {
                     statement.getInt(0)
                 } else {
@@ -45,7 +46,7 @@ fun SQLiteConnection.validateDatabase(tag: String = "", tableDataCountCheck: Str
 
                 if (count == null || (!allowZeroCount && count == 0)) {
                     Logger.e { "validateDatabase - table [$tableDataCountCheck] is BLANK for database [$tag] is blank" }
-                    return false
+                    return@usePrepared false
                 }
             }
         }
@@ -63,7 +64,7 @@ fun SQLiteConnection.validateDatabase(tag: String = "", tableDataCountCheck: Str
  * @param toDatabasePath Path to attach database
  * @param toDatabaseName Alias name for attached database
  */
-fun SQLiteConnection.attachDatabase(toDatabasePath: String, toDatabaseName: String) {
+suspend fun Transactor.attachDatabase(toDatabasePath: String, toDatabaseName: String) {
     val sql = "ATTACH DATABASE '$toDatabasePath' AS $toDatabaseName"
     execSQL(sql)
 }
@@ -72,7 +73,7 @@ fun SQLiteConnection.attachDatabase(toDatabasePath: String, toDatabaseName: Stri
  * Detach a database
  * @param databaseName Alias name for attached database
  */
-fun SQLiteConnection.detachDatabase(databaseName: String) {
+suspend fun Transactor.detachDatabase(databaseName: String) {
     val sql = "DETACH DATABASE '$databaseName'"
     execSQL(sql)
 }
@@ -82,37 +83,39 @@ fun SQLiteConnection.detachDatabase(databaseName: String) {
  *
  * @return List of AttachedDatabaseInfo containing the database name and file path
  */
-fun SQLiteConnection.getAttachedDatabases(): List<AttachedDatabaseInfo> {
-    val statement = prepare("SELECT * FROM pragma_database_list")
-
-    val databaseInfoList = mutableListOf<AttachedDatabaseInfo>()
-    val nameColumnIndex = getColumnIndexOrThrow(statement, "name")
-    val fileColumnIndex = getColumnIndexOrThrow(statement, "file")
-    while (statement.step()) {
-        val name = statement.getText(nameColumnIndex)
-        val file = statement.getText(fileColumnIndex)
-        databaseInfoList.add(AttachedDatabaseInfo(name, file))
+suspend fun Transactor.getAttachedDatabases(): List<AttachedDatabaseInfo> {
+    usePrepared("SELECT * FROM pragma_database_list") { statement ->
+        val databaseInfoList = mutableListOf<AttachedDatabaseInfo>()
+        val nameColumnIndex = getColumnIndexOrThrow(statement, "name")
+        val fileColumnIndex = getColumnIndexOrThrow(statement, "file")
+        while (statement.step()) {
+            val name = statement.getText(nameColumnIndex)
+            val file = statement.getText(fileColumnIndex)
+            databaseInfoList.add(AttachedDatabaseInfo(name, file))
+        }
+        return@usePrepared databaseInfoList
     }
 
-    return databaseInfoList
+    return emptyList()
 }
 
 /**
  * Find names of tables in this database
  * @param databaseName Alias name for database (such as an attached database) (optional)
  */
-fun SQLiteConnection.findTableNames(databaseName: String = ""): List<String> {
-    val statement = if (databaseName.isNotBlank()) {
-        prepare("SELECT tbl_name FROM $databaseName.sqlite_master where type='table'")
-    } else {
-        prepare("SELECT tbl_name FROM sqlite_master where type='table'")
-    }
-
-
+suspend fun Transactor.findTableNames(databaseName: String = ""): List<String> {
     val tableNames = mutableListOf<String>()
-    statement.use {
-        while (it.step()) {
-            tableNames.add(statement.getText(0))
+    if (databaseName.isNotBlank()) {
+        usePrepared("SELECT tbl_name FROM $databaseName.sqlite_master where type='table'") {
+            while (it.step()) {
+                tableNames.add(it.getText(0))
+            }
+        }
+    } else {
+        usePrepared("SELECT tbl_name FROM sqlite_master where type='table'")  {
+            while (it.step()) {
+                tableNames.add(it.getText(0))
+            }
         }
     }
 
@@ -127,7 +130,7 @@ fun SQLiteConnection.findTableNames(databaseName: String = ""): List<String> {
  *
  * @return true If tableName exist
  */
-fun SQLiteConnection.tableExists(tableName: String, databaseName: String = ""): Boolean {
+suspend fun Transactor.tableExists(tableName: String, databaseName: String = ""): Boolean {
     return tablesExists(listOf(tableName), databaseName)
 }
 
@@ -138,20 +141,13 @@ fun SQLiteConnection.tableExists(tableName: String, databaseName: String = ""): 
  *
  * @return true If ALL tableNames exist
  */
-fun SQLiteConnection.tablesExists(tableNames: List<String>, databaseName: String = ""): Boolean {
+suspend fun Transactor.tablesExists(tableNames: List<String>, databaseName: String = ""): Boolean {
     val inClaus = tableNames.joinToString(",", prefix = "(", postfix = ")") { "'$it'" }
 
-    val statement = if (databaseName.isNotBlank()) {
-        prepare("SELECT count(1) FROM $databaseName.sqlite_master WHERE type='table' AND tbl_name IN $inClaus")
+    val tableCount = if (databaseName.isNotBlank()) {
+        execIntResultSql("SELECT count(1) FROM $databaseName.sqlite_master WHERE type='table' AND tbl_name IN $inClaus")
     } else {
-        prepare("SELECT count(1) FROM sqlite_master WHERE type='table' AND tbl_name IN $inClaus")
-    }
-
-    var tableCount = 0
-    statement.use {
-        while (it.step()) {
-            tableCount = it.getInt(0)
-        }
+        execIntResultSql("SELECT count(1) FROM sqlite_master WHERE type='table' AND tbl_name IN $inClaus")
     }
 
     return tableCount == tableNames.size
@@ -161,17 +157,19 @@ fun SQLiteConnection.tablesExists(tableNames: List<String>, databaseName: String
  * Find names of views in this database
  * @param databaseName Alias name for database (such as an attached database) (optional)
  */
-fun SQLiteConnection.findViewNames(databaseName: String = ""): List<String> {
-    val statement = if (databaseName.isNotBlank()) {
-        prepare("SELECT tbl_name FROM $databaseName.sqlite_master where type='view'")
-    } else {
-        prepare("SELECT tbl_name FROM sqlite_master where type='view'")
-    }
-
+suspend fun Transactor.findViewNames(databaseName: String = ""): List<String> {
     val viewNames = mutableListOf<String>()
-    statement.use {
-        while (it.step()) {
-            viewNames.add(it.getText(0))
+    if (databaseName.isNotBlank()) {
+        usePrepared("SELECT tbl_name FROM $databaseName.sqlite_master where type='view'") {
+            while (it.step()) {
+                viewNames.add(it.getText(0))
+            }
+        }
+    } else {
+        usePrepared("SELECT tbl_name FROM sqlite_master where type='view'") {
+            while (it.step()) {
+                viewNames.add(it.getText(0))
+            }
         }
     }
 
@@ -186,7 +184,7 @@ fun SQLiteConnection.findViewNames(databaseName: String = ""): List<String> {
  *
  * @return true If viewName exist
  */
-fun SQLiteConnection.viewExists(viewName: String, databaseName: String = ""): Boolean {
+suspend fun Transactor.viewExists(viewName: String, databaseName: String = ""): Boolean {
     return viewExists(listOf(viewName), databaseName)
 }
 
@@ -197,29 +195,22 @@ fun SQLiteConnection.viewExists(viewName: String, databaseName: String = ""): Bo
  *
  * @return true If ALL viewNames exist
  */
-fun SQLiteConnection.viewExists(viewNames: List<String>, databaseName: String = ""): Boolean {
+suspend fun Transactor.viewExists(viewNames: List<String>, databaseName: String = ""): Boolean {
     val inClaus = viewNames.joinToString(",", prefix = "(", postfix = ")") { "'$it'" }
 
-    val statement = if (databaseName.isNotBlank()) {
-        prepare("SELECT count(1) FROM $databaseName.sqlite_master WHERE type='view' AND tbl_name IN $inClaus")
+    val viewCount = if (databaseName.isNotBlank()) {
+        execIntResultSql("SELECT count(1) FROM $databaseName.sqlite_master WHERE type='view' AND tbl_name IN $inClaus")
     } else {
-        prepare("SELECT count(1) FROM sqlite_master WHERE type='view' AND tbl_name IN $inClaus")
-    }
-
-    var viewCount = 0
-    statement.use {
-        while (it.step()) {
-            viewCount = it.getInt(0)
-        }
+        execIntResultSql("SELECT count(1) FROM sqlite_master WHERE type='view' AND tbl_name IN $inClaus")
     }
 
     return viewCount == viewNames.size
 }
 
-internal fun SQLiteConnection.execIntResultSql(sql: String, columnIndex: Int = 0): Int? {
-    this.prepare(sql).use { statement ->
+internal suspend fun Transactor.execIntResultSql(sql: String, columnIndex: Int = 0): Int? {
+    this.usePrepared(sql) { statement ->
         if (statement.step()) {
-            return statement.getInt(columnIndex)
+            return@usePrepared statement.getInt(columnIndex)
         } else {
             Logger.w { "Failed to get Int for [$sql] (returned NO data)" }
         }
@@ -228,10 +219,10 @@ internal fun SQLiteConnection.execIntResultSql(sql: String, columnIndex: Int = 0
     return null
 }
 
-internal fun SQLiteConnection.execTextResultSql(sql: String, columnIndex: Int = 0): String? {
-    this.prepare(sql).use { statement ->
+internal suspend fun Transactor.execTextResultSql(sql: String, columnIndex: Int = 0): String? {
+    this.usePrepared(sql) { statement ->
         if (statement.step()) {
-            return statement.getText(columnIndex)
+            return@usePrepared statement.getText(columnIndex)
         } else {
             Logger.w { "Failed to get Text for [$sql] (returned NO data)" }
         }
@@ -245,7 +236,7 @@ internal fun SQLiteConnection.execTextResultSql(sql: String, columnIndex: Int = 
  *
  * @param viewName Name of view to drop
  */
-fun SQLiteConnection.dropView(viewName: String) {
+suspend fun Transactor.dropView(viewName: String) {
     execSQL("DROP VIEW IF EXISTS $viewName")
 }
 
@@ -254,20 +245,20 @@ fun SQLiteConnection.dropView(viewName: String) {
  *
  * @param views List of view names... if this list is empty then all views in the database will be dropped
  */
-fun SQLiteConnection.dropAllViews(views: List<String> = emptyList()) {
+suspend fun Transactor.dropAllViews(views: List<String> = emptyList()) {
     val viewNames: List<String> = views.ifEmpty { findViewNames() }
     viewNames.forEach { dropView(it) }
 }
 
-fun SQLiteConnection.createView(viewName: String, viewQuery: String) {
+suspend fun Transactor.createView(viewName: String, viewQuery: String) {
     execSQL("CREATE VIEW `$viewName` AS $viewQuery")
 }
 
-fun SQLiteConnection.createAllViews(views: List<DatabaseViewQuery>) {
+suspend fun Transactor.createAllViews(views: List<DatabaseViewQuery>) {
     views.forEach { createView(it.viewName, it.viewQuery.trim()) }
 }
 
-fun SQLiteConnection.recreateView(viewName: String, viewQuery: String) {
+suspend fun Transactor.recreateView(viewName: String, viewQuery: String) {
     dropView(viewName)
     createView(viewName, viewQuery)
 }
@@ -277,7 +268,7 @@ fun SQLiteConnection.recreateView(viewName: String, viewQuery: String) {
  *
  * @param views List of Views to recreate
  */
-fun SQLiteConnection.recreateAllViews(views: List<DatabaseViewQuery>) {
+suspend fun Transactor.recreateAllViews(views: List<DatabaseViewQuery>) {
     views.forEach { recreateView(it.viewName, it.viewQuery.trim()) }
 }
 
@@ -290,18 +281,18 @@ fun SQLiteConnection.recreateAllViews(views: List<DatabaseViewQuery>) {
  *
  * @return true If all SQL statements successfully were applied
  */
-fun SQLiteConnection.applySqlFile(fileSystem: FileSystem, sqlPath: Path): Boolean {
+fun Transactor.applySqlFile(fileSystem: FileSystem, sqlPath: Path): Boolean {
     if (!fileSystem.exists(sqlPath)) {
         // Can't apply if there is no file
         Logger.e { "Failed to apply sql file. File: [$sqlPath] does NOT exist" }
         return false
     }
-    
-    return runInTransaction {
-        fileSystem.parseAndExecuteSqlStatements(sqlPath) { statement ->
-            this.execSQL(statement)
-        }
+
+    fileSystem.parseAndExecuteSqlStatements(sqlPath) { statement ->
+        runBlocking { this@applySqlFile.execSQL(statement) }
     }
+
+    return true
 }
 
 /**
@@ -311,7 +302,7 @@ fun SQLiteConnection.applySqlFile(fileSystem: FileSystem, sqlPath: Path): Boolea
  * @param alterSql SQL to be run if the column does not exist.
  * Example: alterTableIfColumnDoesNotExist(database, "individual", "middle_name", "ALTER TABLE individual ADD `middle_name` TEXT DEFAULT '' NOT NULL")
  */
-fun SQLiteConnection.alterTableIfColumnDoesNotExist(tableName: String, columnName: String, alterSql: String) {
+suspend fun Transactor.alterTableIfColumnDoesNotExist(tableName: String, columnName: String, alterSql: String) {
     if (!this.columnExists(tableName, columnName)) {
         Logger.i { "Adding column [$columnName] to table [$tableName]" }
         execSQL(alterSql)
@@ -326,10 +317,10 @@ fun SQLiteConnection.alterTableIfColumnDoesNotExist(tableName: String, columnNam
  * @return true if the column exists otherwise false
  */
 @Suppress("NestedBlockDepth")
-fun SQLiteConnection.columnExists(tableName: String, columnName: String): Boolean {
+suspend fun Transactor.columnExists(tableName: String, columnName: String): Boolean {
     var columnExists = false
 
-    this.prepare("PRAGMA table_info($tableName)").use { statement: SQLiteStatement ->
+    this.usePrepared("PRAGMA table_info($tableName)") { statement: SQLiteStatement ->
         if (statement.step()) {
             do {
                 val currentColumn = statement.getText(statement.getColumnIndexOrThrow("name"))
@@ -351,7 +342,7 @@ fun SQLiteConnection.columnExists(tableName: String, columnName: String): Boolea
  *
  * @return version of the database
  */
-fun SQLiteConnection.getDatabaseVersion(): Int {
+suspend fun Transactor.getDatabaseVersion(): Int {
     return execIntResultSql("PRAGMA user_version") ?: 0
 }
 
@@ -360,7 +351,7 @@ fun SQLiteConnection.getDatabaseVersion(): Int {
  *
  * @param newVersion version to be set on database
  */
-fun SQLiteConnection.setDatabaseVersion(newVersion: Int) {
+suspend fun Transactor.setDatabaseVersion(newVersion: Int) {
     execSQL("PRAGMA user_version = $newVersion")
 }
 
@@ -369,39 +360,18 @@ fun SQLiteConnection.setDatabaseVersion(newVersion: Int) {
  *
  * @param newVersion version to be set on database (default to 0)
  */
-fun SQLiteConnection.resetRoom(newVersion: Int = 0) {
+suspend fun Transactor.resetRoom(newVersion: Int = 0) {
     execSQL("DROP TABLE IF EXISTS room_master_table")
     setDatabaseVersion(newVersion)
-}
-
-/**
- * Begin Transaction
- */
-fun SQLiteConnection.beginTransaction() {
-    execSQL("BEGIN IMMEDIATE TRANSACTION")
-}
-
-/**
- * End Transaction
- */
-fun SQLiteConnection.endTransaction() {
-    execSQL("END TRANSACTION")
-}
-
-/**
- * Rollback Transaction
- */
-fun SQLiteConnection.rollbackTransaction() {
-    execSQL("ROLLBACK TRANSACTION")
 }
 
 /**
  * Check integrity of a database connection using PRAGMA command
  * @return true if integrity_check == "ok"
  */
-fun SQLiteConnection.isIntegrityOk(): Boolean {
+suspend fun Transactor.isIntegrityOk(): Boolean {
     var resultText: String? = ""
-    this.prepare("PRAGMA integrity_check").use { statement: SQLiteStatement ->
+    this.usePrepared("PRAGMA integrity_check") { statement: SQLiteStatement ->
         resultText = if (statement.step()) {
             statement.getText(0)
         } else {
@@ -423,24 +393,6 @@ fun SQLiteConnection.isIntegrityOk(): Boolean {
 }
 
 /**
- * Executes the specified block in a database transaction. The transaction will be
- * marked as successful unless an exception is thrown in the block.
- */
-@Suppress("TooGenericExceptionCaught")
-inline fun SQLiteConnection.runInTransaction(block: () -> Unit): Boolean {
-    beginTransaction()
-    return try {
-        block()
-        endTransaction()
-        true
-    } catch (transactionException: Throwable) {
-        Logger.i(transactionException) { "Failed to execute transaction.  Message: ${transactionException.message}" }
-        rollbackTransaction()
-        false
-    }
-}
-
-/**
  * If the database should NOT have a migration and is a pre-populated database that should not be managed by Room... make sure Room migration is never needed.
  *
  * NOTE: this SHOULD be called BEFORE room has a chance to open the database and verify the database
@@ -458,7 +410,7 @@ inline fun SQLiteConnection.runInTransaction(block: () -> Unit): Boolean {
  * @param expectedVersion SQLite Database version (PRAGMA user_version)
  * @param expectedIdentityHash Hash that is expected.  If the expectedIdentityHash does not match the existing identity hash (currently in the room_master_table), then just delete the table
  */
-fun SQLiteConnection.checkAndFixRoomIdentityHash(expectedVersion: Int, expectedIdentityHash: String) {
+suspend fun Transactor.checkAndFixRoomIdentityHash(expectedVersion: Int, expectedIdentityHash: String) {
     if (expectedIdentityHash.isBlank()) {
         Logger.e { "checkAndFixRoomIdentityHash -- expectedIdentityHash is blank" }
         return
@@ -476,10 +428,8 @@ fun SQLiteConnection.checkAndFixRoomIdentityHash(expectedVersion: Int, expectedI
     }
 
     Logger.w { "checkAndFixRoomIdentityHash -- updating expectedIdentityHash: [$expectedIdentityHash]" }
-    runInTransaction {
-        execSQL("CREATE TABLE IF NOT EXISTS ${Room.MASTER_TABLE_NAME} (id INTEGER PRIMARY KEY,identity_hash TEXT)")
-        execSQL("INSERT OR REPLACE INTO ${Room.MASTER_TABLE_NAME} (id,identity_hash) VALUES(42, '$expectedIdentityHash')")
-    }
+    execSQL("CREATE TABLE IF NOT EXISTS ${Room.MASTER_TABLE_NAME} (id INTEGER PRIMARY KEY,identity_hash TEXT)")
+    execSQL("INSERT OR REPLACE INTO ${Room.MASTER_TABLE_NAME} (id,identity_hash) VALUES(42, '$expectedIdentityHash')")
 }
 
 /**
@@ -488,14 +438,7 @@ fun SQLiteConnection.checkAndFixRoomIdentityHash(expectedVersion: Int, expectedI
  *
  * @return identity_hash for this database OR null if it does exist
  */
-fun SQLiteConnection.findRoomIdentityHash(): String? {
+suspend fun Transactor.findRoomIdentityHash(): String? {
     // NOTE: the id column for this table always seems to be 42 and there is always only 1 row... so lets just find the first row
     return execTextResultSql("SELECT identity_hash FROM room_master_table LIMIT 1")
-}
-
-fun SQLiteConnection.execSQLWithArgs(sql: String, args: List<Any?>) {
-    this.prepare(sql).use { statement ->
-        statement.bindArgs(args)
-        statement.step() // Execute the statement
-    }
 }